+++ releasenotes.md
@@ -3,3 +3,4 @@ Whats New:
 - A rating option has been added so each formula can be rated
 - Added the option to attach documents in formulas
 - Formula making view improvements
+- Generate ingredient SDS from the available information